pyproject.toml
[tool.poetry]
name = "article-scraper"
version = "0.1.0"
description = "A simple scraper using scrapy"
authors = ["Alexander Price"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.6.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
